Add HostInfo to CrashLog
Keep CrashLog version because it only add stuff and do not break back compatibility
